6.1 - Check types and move them to required folders.
