+++ src/entities/album/api/albumsApi.ts
@@ -1,6 +1,6 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { API_URL } from '../../../config/api-config.ts';
-import type { AlbumProps } from '../types/types.ts';
+import type { AlbumProps } from '../model/types.ts';
 
 export const albumsApi = createApi({
   reducerPath: 'albumsApi',
