+++ src/entities/comment/model/types.ts
