+++ src/entities/photo/model/types.ts
