+++ src/entities/photo/ui/Photo.tsx
@@ -1,5 +1,5 @@
 import styles from './Photo.module.scss';
-import type { PhotoProps } from '../types/types.ts';
+import type { PhotoProps } from '../model/types.ts';
 import React, { type FC, useState } from 'react';
 
 type PhotoComponentProps = {
