+++ src/entities/post/model/slice/postSlice.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter, createSlice } from '@reduxjs/toolkit';
-import type { PostProps } from '../../types/types.ts';
+import type { PostProps } from '../types.ts';
 import { postsApi } from '../../api/postsApi.ts';
 import type { AppState } from '../../../../app/providers/store/store.ts';
